# Generate debug symbols in debug builds only