@bkaradzic 2020-11-09
This is intentional... Symbols are stripped from executable, but libraries are containing them:

Executables:
```
Before:                                        | After:
2.5M Nov  8 23:25 example-17-drawstressRelease | 2.5M Nov  8 23:18 example-17-drawstressRelease
923K Nov  8 23:25 example-25-c99Release        | 923K Nov  8 23:18 example-25-c99Release
3.3M Nov  8 23:25 examplesRelease              | 3.3M Nov  8 23:18 examplesRelease
331K Nov  8 23:25 geometrycRelease             | 331K Nov  8 23:18 geometrycRelease
2.7M Nov  8 23:25 geometryvRelease             | 2.7M Nov  8 23:18 geometryvRelease
```